+++ example/dartway_example_flutter/lib/app/news/news_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:dartway_example_flutter/app/news/widgets/create_news_post_fab.dart';
 import 'package:dartway_example_flutter/app/news/widgets/news_post_list.dart';
-import 'package:dartway_example_flutter/common/app_scaffold.dart';
+import 'package:dartway_example_flutter/shared/widgets/app_scaffold.dart';
 import 'package:dartway_example_flutter/core/app_l10n.dart';
 import 'package:dartway_example_flutter/core/user_profile_provider.dart';
 import 'package:dartway_example_flutter/core/user_profile_roles.dart';