+++ example/dartway_example_flutter/lib/app/schedule/schedule_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:dartway_example_flutter/app/schedule/widgets/schedule_app_bar.dart';
 import 'package:dartway_example_flutter/app/schedule/widgets/schedule_session_list.dart';
-import 'package:dartway_example_flutter/common/app_scaffold.dart';
+import 'package:dartway_example_flutter/shared/widgets/app_scaffold.dart';
 import 'package:dartway_example_flutter/ui_kit/ui_kit.dart';
 
 class SchedulePage extends ConsumerWidget implements DwFeature {